# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the BuffetPOS API documentation.",
        "title": "BuffetPOS API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/customer/invoices": {
            "get": {
                "description": "Get invoice by table ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Get Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Code",
                        "name": "AccessCode",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.InvoiceDetail"
                        }
                    }
                }
            }
        },
...
        "/manage/invoices/charge-food-overweight": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Charge Fee Food Overweight By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Charge Fee Food Overweight",
                "parameters": [
                    {
                        "description": "Charge Fee Food Overweight Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.ChargeFeeFoodOverWeightRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/manage/invoices/paid": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all paid invoices.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get All Paid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.InvoiceDetail"
                            }
                        }
                    }
                }
            }
        },
        "/manage/invoices/set-paid": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Set invoice as paid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Set Invoice Paid",
                "parameters": [
                    {
                        "description": "Update Invoice Status Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateInvoiceStatusRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/manage/invoices/unpaid": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all unpaid invoices.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get All Unpaid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.InvoiceDetail"
                            }
                        }
                    }
                }
            }
        },
        "/manage/invoices/{invoice_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancel invoice.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Cancel Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "invoice_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.ChargeFeeFoodOverWeightRequest": {
            "type": "object",
            "required": [
                "invoice_id",
                "total_food_weight"
            ],
            "properties": {
                "invoice_id": {
                    "type": "string"
                },
                "total_food_weight": {
                    "type": "number"
                }
            }
        },
...
        "requests.UpdateInvoiceStatusRequest": {
            "type": "object",
            "required": [
                "invoice_id"
            ],
            "properties": {
                "invoice_id": {
                    "type": "string"
                }
            }
        },
...
        "responses.InvoiceDetail": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customer_phone": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isPaid": {
                    "type": "boolean"
                },
                "peopleAmount": {
                    "type": "integer"
                },
                "price_fee_food_overweight": {
                    "type": "number"
                },
                "tableId": {
                    "type": "string"
                },
                "totalPrice": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}